+++ src/components/Common/Icons/Pause.tsx
