+++ src/components/Common/Icons/SpeakerWave.tsx
